Chore: domain 모듈의 dto 패키지를 각 모듈로 이동 (#126)
* Refactor: 커피챗 수정 메서드 인자를 엔티티 클래스로 래핑하도록 변경

* Chore: domain 모듈의 dto패키지를 각 모듈로 이동

* Chore: Q타입 객체 삭제
