+++ module-api/src/main/java/kernel/jdon/coffeechat/dto/response/CreateCoffeeChatResponse.java
@@ -1,9 +1,7 @@
 package kernel.jdon.coffeechat.dto.response;
 
-import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
 @AllArgsConstructor
