+++ module-api/src/main/java/kernel/jdon/coffeechat/dto/response/UpdateCoffeeChatResponse.java
@@ -1,12 +1,6 @@
 package kernel.jdon.coffeechat.dto.response;
 
-import java.time.LocalDateTime;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-
-import kernel.jdon.coffeechat.domain.CoffeeChat;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 
 @Getter
