+++ module-api/src/main/java/kernel/jdon/member/dto/response/RemoveMemberResponse.java
