+++ module-api/src/main/java/kernel/jdon/skill/dto/response/FindJobCategorySkillResponse.java
@@ -1,6 +1,4 @@
-package kernel.jdon.skill.dto;
-
-import java.util.List;
+package kernel.jdon.skill.dto.response;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -12,7 +10,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor
-public class FindListDataBySkillResponse {
-	private List<FindLectureResponse> lectureList;
-	private List<FindJdResponse> jdList;
+public class FindJobCategorySkillResponse {
+	private Long skillId;
+	private String keyword;
 }
