+++ module-api/src/main/java/kernel/jdon/skill/dto/response/FindLectureResponse.java
@@ -1,0 +1,22 @@
+package kernel.jdon.skill.dto.response;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+@AllArgsConstructor
+public class FindLectureResponse {
+	private Long lectureId;
+	private String title;
+	private String lectureUrl;
+	private String imageUrl;
+	private String instructor;
+	private Integer studentCount;
+	private Integer price;
+	private Boolean isFavorite;
+}
