+++ module-api/src/main/java/kernel/jdon/skill/dto/response/FindListMemberSkillResponse.java
@@ -1,4 +1,6 @@
-package kernel.jdon.skill.dto;
+package kernel.jdon.skill.dto.response;
+
+import java.util.List;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -10,7 +12,6 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor
-public class FindHotSkillResponse {
-	private Long skillId;
-	private String keyword;
+public class FindListMemberSkillResponse {
+	private List<FindMemberSkillResponse> skillList;
 }
